tests: Convert snapshot tests to integration tests
This moves the snapshot tests from the `tests` module in `src/lib.rs` into
`tests/generate.rs` so they exercise the crate through its public API.

The `format_bytes()`/`format_number()` tests stay as unit tests since
they cover private functions.
