+++ tests/snapshots/generate__404-avatar.snap.png
