+++ tests/snapshots/generate__asian-text-description.snap.png
