+++ tests/snapshots/generate__generated_og_image.snap.png
