+++ tests/snapshots/generate__generated_og_image_escaping.snap.png
