+++ tests/snapshots/generate__generated_og_image_minimal.snap.png
