+++ tests/snapshots/generate__generated_og_image_overflow.snap.png
