+++ tests/snapshots/generate__unicode-truncation.snap.png
