+++ tests/snapshots/generate__whitespace-handling.snap.png
